feat: better Address types
