+++ packages/address/Address.stories.tsx
@@ -1,6 +1,6 @@
+import React from 'react'
 import { Story, Meta } from '@storybook/react'
-import { AddressProps } from './types'
-import Address from './Address'
+import { Address, AddressProps } from '.'
 
 export default {
   component: Address,
