+++ packages/address/index.ts
@@ -1,3 +1,2 @@
 export { default as Address } from './Address'
-export * from './trimAddress'
-export * from './types'
+export * from './Address'
