--- packages/address/types.ts
@@ -1,9 +1,0 @@
-import { LidoComponentProps } from '@lidofinance/utils'
-
-export type AddressProps = LidoComponentProps<
-  'div',
-  {
-    address: string
-    symbols?: number
-  }
->
